Implemented offline discussion
